chore: move load_from_yaml in to walrus_utils
