+++ crates/walrus-utils/src/lib.rs
@@ -1,6 +1,10 @@
 // Copyright (c) Walrus Foundation
 // SPDX-License-Identifier: Apache-2.0
 
+use std::path::Path;
+
+use serde::de::DeserializeOwned;
+
 #[cfg(feature = "backoff")]
 pub mod backoff;
 
@@ -24,6 +28,23 @@ pub mod tests {
         static LOCK: OnceLock<Mutex<()>> = OnceLock::new();
         LOCK.get_or_init(Mutex::default)
     }
+}
+
+/// Load the config from a YAML file located at the provided path.
+pub fn load_from_yaml<P: AsRef<Path>, T: DeserializeOwned>(path: P) -> anyhow::Result<T> {
+    use anyhow::Context;
+
+    let path = path.as_ref();
+    tracing::debug!(path = %path.display(), "[load_from_yaml] reading from file");
+
+    let reader = std::fs::File::open(path).with_context(|| {
+        format!(
+            "[load_from_yaml] unable to load config from {}",
+            path.display()
+        )
+    })?;
+
+    Ok(serde_yaml::from_reader(reader)?)
 }
 
 /// A macro to print a crumb of information to the console. This is useful for debugging.
